requirements.txt
torch
pillow
torchvision